requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
groq==0.9.0
pdfplumber==0.11.0
numpy==1.26.4
pydantic==2.9.0
python-multipart==0.0.9
python-dotenv==1.0.0